requirements.txt
Flask
pymongo
gridfs
python-dotenv
pandas
google-generativeai
Werkzeug
bson
